PlotDarkMatter {Vennerable}	R Documentation

Plots dark matter

Description

Fills the area of a VennDrawing universe that is not occupied by any set.

Usage

PlotDarkMatter(VD)

Arguments

VD
An object of class VennDrawing

Details

This works by filling the entire universe with grey (not currently changeable) and then plotting then filling the inside of the dark matter boundary with white.

Value

Executed for its side effecrs

Author(s)

Jonathan Swinton (jonathan@swintons.net)

Examples

VD <- compute.Venn(Venn(n=2))
PlotDarkMatter(VD)

[Package Vennerable version 3.1.0.9000 Index]