PlotVennGeometry {Vennerable}	R Documentation

Display computed Venn diagrams

Description

This is the function called by the plot generic for objects of class VennDrawing.

Usage

PlotVennGeometry(C3, gpList, show = list(FaceText = "weight"))

Arguments

C3
An object of class VennDrawing, as produced by compute.Venn

gpList
A list of lists of graphical parameters. The VennThemes function produces a suitable list; see its man page for more details.

show
A list of options controlling what is shown. Defaults are universe = TRUE, Sets = TRUE, SetLabels = TRUE, DarkMatter = FALSE, Faces = TRUE, FaceText = "weight". If universe = TRUE the bounding region, currently always a rectangle, of the diagram is shown. If Sets = TRUE, the boundaries of the Sets are shown. If SetLabels = TRUE the names of the Sets are plotted according to the instructions in VennGetSetLabels(C3). If Faces = TRUE individual faces are filled. FaceText is a character vector which may contain any of c("weight","signature","sets","elements"). Currently DarkMatter must be FALSE.

Details

The algorithm for displaying face annotation is pretty crude. The annotation corresponding to each member of the FaceText vector is displayed on a single line. weight and signature display the associated weight and signature for the face. sets produces a concatenated version of the names of the sets in the face, while elements does the same for the elements; both of these only really work for single character set names or elements.

Author(s)

Jonathan Swinton (jonathan@swintons.net)

Examples

library(Vennerable)
setList <- strsplit(month.name,split="")
names(setList) <- month.name
C3 <- compute.Venn(Venn(setList[1:3]))
PlotVennGeometry(C3,show=list(FaceText="signature"),gp=VennThemes(C3,colourAlgorithm="signature"))

[Package Vennerable version 3.1.0.9000 Index]