StemCell {Vennerable}	R Documentation

Stem cell data

Description

Data on the genes which bind to each of three transcription factors (and a fourth which may also be a transcription factor)

Usage

data(StemCell)

Format

A named list with four elements named OCT4, SOX2, NANOG and E2F4. Each element is a character vector containing the names of 600-1700 associated genes.

Source

Taken from supplementary information in Boyer et al, "Core transcriptional regulatory circuitry in human embryonic stem cells", Cell, 2005, 122:947–956, 10.1016/j.cell.2005.08.020,http://www.ncbi.nlm.nih.gov/pubmed/16153702.

Examples

library(Vennerable)
data(StemCell)
plot(Venn(Sets=StemCell),type="squares",doWeights=FALSE)

[Package Vennerable version 3.1.0.9000 Index]