VennDiagrams {Vennerable}	R Documentation

Predrawn Venn diagrams.

Description

Precomputed data on the geometrical structure of various types of Venn diagrams.

Usage

data(VennDiagrams)

Format

A list with elements named battle,AWFEscale,AWFE, and ellipses. Each element is itself a list in which the nth element, if not NULL, is an object of class TissueDrawing representing an example of an unweighted Venn diagram on n sets of the given type.

Details

This is really a device to cache the diagrams and is not for direct use. Use eg compute.Venn(Venn(n=...),type=... which returns a plottable object of class VennDrawing instead.

Source

Created by the buildVennDiagrams() function within Vennerable

Examples

data(VennDiagrams)
TD <- VennDiagrams[["ellipses"]][[4]]
class(TD)

[Package Vennerable version 3.1.0.9000 Index]