VennDrawing-class {Vennerable}	R Documentation

Class "VennDrawing"

Description

A geometrical representation of a Venn diagram suitable for display

Objects from the Class

Objects can be created by calls of the form new("VennDrawing",V,TD) where V is an object of class Venn and TD is an object of class TissueDrawing which has representations for all the Sets and Faces corresponding to the sets and intersections of the V object.

As well as bringing together the geometric and set-theoretic properties of the Venn diagram, this class stores information about the universe in which to display the diagram and provides a number of methods for plotting the final figure.

Slots

universe:

A matrix with the x and y coordinates of the bounding universe. Currently assumed to have 2 rows by much of the code, corresponding to the lower left and upper right corners of a bounding rectangle

SetLabels:

A data.frame encoding the positions of labels for the Sets

FaceLabels:

A data.frame encoding the positions of labels for the Faces

setList:

See TissueDrawing-class

nodeList:

See TissueDrawing-class

recentChanges:

See TissueDrawing-class

edgeList:

See TissueDrawing-class

faceList:

See TissueDrawing-class

faceSignature:

See TissueDrawing-class

IndicatorWeight:

See Venn-class

IntersectionSets:

See Venn-class

Extends

Class "TissueDrawing", directly. Class "Venn", directly. Class "TDEdgeList", by class "TissueDrawing", distance 2. Class "TDFaceList", by class "TissueDrawing", distance 2.

Methods

Areas

signature(object = "VennDrawing"): Returns the area of each named face

IntersectionMidpoints

signature(object = "VennDrawing"): ...

plot

signature(x = "VennDrawing", y = "missing"): ...

PlotUniverse

signature(object = "VennDrawing"): ...

show

signature(object = "VennDrawing"): ...

UniverseRange

signature(object = "VennDrawing"): ...

VisibleRange

signature(object = "VennDrawing"): ...

Author(s)

Jonathan Swinton (jonathan@swintons.net)

Examples

showClass("VennDrawing")

[Package Vennerable version 3.1.0.9000 Index]