VennSetSetLabels {Vennerable}	R Documentation

Set and get annotation labels and graphical styles

Description

Embedded within a VennDiagram object are instructions for displaying annotation for the sets and their faces. Retrieve them so they can be edited and re-embedded for subsequent display.

Usage

VennGetSetLabels(object)
VennSetSetLabels(object, SetLabels)
VennGetFaceLabels(object)
VennSetFaceLabels(object, FaceLabels)
VennGetUniverseRange(object)
VennSetUniverseRange(object, universe)

Arguments

object
An object of class VennDrawing

SetLabels
A data.frame. The format of this data.frame may change.

FaceLabels
A data.frame. The format of this data.frame may change.

universe
An nx2 matrix defining x and y coordinates of the bounding universe. Currently assumed to have 2 rows by much of the code, corresponding to the lower left and upper right corners of a bounding rectangle

Author(s)

Jonathan Swinton (jonathan@swintons.net)

Examples

data(StemCell)
Vstem <- Venn(StemCell)
Vstem3 <- Vstem[,c("OCT4","SOX2","NANOG")]
Cstem3 <- compute.Venn(Vstem3,doWeights=TRUE)
plot(Cstem3)
# don't like the default position of 'SOX2'?
SetLabels <- VennGetSetLabels(Cstem3)
SetLabels[SetLabels$Label=="SOX2","x"] <- 12
Cstem3 <- VennSetSetLabels(Cstem3,SetLabels)
grid::grid.newpage()
plot(Cstem3)

[Package Vennerable version 3.1.0.9000 Index]