compute.S4 {Vennerable}	R Documentation

Venn diagrams on 4 sets

Description

Given a Venn object on 4 sets, returns an VennDrawing object rendering it in the Lewis Carroll style.

Usage

compute.S4(V, doWeights = FALSE, s = 0.25, likeSquares = TRUE)

Arguments

V
A Venn object on 4 sets.

doWeights
Must be FALSE. For weighted diagrams, use compute.CR.

s
If s=0 then set boundaries may be superimposed; for s>0 the larger two rectangles are offset by s.

likeSquares
If TRUE, make the individual faces as square as possible; if FALSE remove additional corners and use triangular edges.

Details

The placement of the face names is not ideal for some combinations with s>0.

Value

An object of class VennDrawing

Author(s)

Jonathan Swinton (jonathan@swintons.net)

Examples

compute.S4(Venn(n=4))

[Package Vennerable version 3.1.0.9000 Index]